fix: Resolve API endpoint issues and implement database seeding
- Fix Achievement model percentage field from int to float for 99.99% values
- Add comprehensive database connection setup in dependencies.py
- Create multiple database seeder scripts for existing schema
- Populate translation tables with multilingual UI content
- Document all endpoint testing status and known issues
- All static data endpoints now working (require trailing slash)
- Analytics endpoints working with fallback data
- Database successfully seeded with basic content

+++ backend/app/models/portfolio.py
@@ -53,7 +53,7 @@ class Achievement(BaseModel):
     description: str
     icon: str
     color: str
-    percentage: int
+    percentage: float
     details: List[str]
     category: str
     created_at: datetime
